Revert "frontend: Mejora ui"
