+++ Frontend/src/styles/gestionPedidosPage.module.css
@@ -35,7 +35,7 @@
   display: flex;
   gap: 10px;
   flex-grow: 1;
-  max-width: 670px;
+  max-width: 600px;
 }
 
 .searchInput::placeholder {
@@ -77,7 +77,7 @@
 .itemInfo {
   display: flex;
   flex-direction: column;
-  gap: 30px;
+  gap: 4px;
   flex-grow: 1;
 }
 .itemInfo strong {
@@ -87,7 +87,6 @@
   font-size: 0.9rem;
   color: #555;
 }
-
 
 .itemActions {
   display: flex;
